### 2020-07-09 #2456 @whedon
[PRE REVIEW]: DrWatson: the perfect sidekick for your scientific inquiries
**Submitting author:** @tamasgal (<a href="http://orcid.org/0000-0001-7821-8673">Tamás Gál</a>)
**Repository:** <a href="https://github.com/JuliaDynamics/DrWatson.jl" target ="_blank">https://github.com/JuliaDynamics/DrWatson.jl</a>
**Version:** v1.13.2
