### 2020-07-09 #2456 @kyleniemeyer
Hi @tamasgal, as we are getting started here, could you please add a Statement of Need section? This is now required for articles, as described in [our documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).

Also, it looks like there is a problem with a DOI ☝️
