### 2020-07-10 116a62b
Slight change to fulfil JOSS requirements
