### 2020-07-10 64013cb
A few fixes to fulfil JOSS requirements (#173)
* Fix DOI

* Slight change to fulfil JOSS requirements

* Update affiliation

Co-authored-by: Sebastian Pech <sebastian.pech@me.com>
