+++ paper/paper.bib
@@ -6,7 +6,7 @@
     pages = "415 - 448",
     year = "2004",
     issn = "0747-7171",
-    doi = "https://doi.org/10.1016/j.jsc.2003.04.002",
+    doi = "10.1016/j.jsc.2003.04.002",
     url = "http://www.sciencedirect.com/science/article/pii/S0747717103001184",
     author = "Andrew V. Sills",
     keywords = "Rogers–Ramanujan identities, -series, Experimental mathematics",
