### 2020-07-13 f1a2392
Update project_setup.jl
setup failed if `placeholder` = true; I think `files` should be `ph_files` here.
