### 2020-07-15 3ae4c2a
cleaned 2 typos (#176)
