### 2020-07-16 b0357ed
improve rationale line
Parameters.jl is kind of obsolete since `Base.@kwdef` eixsts, and this way we highlight more the useful thing, i.e. getting a dataframe of saved results.
