+++ docs/src/index.md
@@ -25,7 +25,7 @@ Have you thought things like:
 * Do I have to produce a dataframe of my finished simulations AGAIN?!
 * Wait, are those experiments already processed?
 * PFfffff I am tired of typing `savename = "w=$w_f=$f_x=$x.txt"`, can't I do it automatically?
-* I wish I could just use Parameters.jl and just translate my simulations into a dataframe.
+* I wish I could just make a dataframe out of all my simulations with one command!
 * Yeah you've sent me your project but none of the scripts work...
 * It would be so nice to automatically integrate `git` information to all the data I save...
 
