### 2020-07-17 4f8f787
mention usage of dataframes in workflow
Closes #177
