+++ docs/src/workflow.jl
@@ -236,6 +236,8 @@ wload(datadir("simulations", firstsim))
 
 # Cool, now we can start analyzing some simulations. The actual analysis is your job,
 # but DrWatson can help you get started with the [`collect_results`](@ref) function.
+# Notice that you need to be `using DataFrames` to access the function!
+using DataFrames
 
 df = collect_results(datadir("simulations"))
 
