### 2020-07-18 #2456 @dpsanders
:wave @tamasgal Sorry for the delay.

Could you please check for each piece of software that you cite whether there is a related published paper. For example, `sumatra` has an associated paper.

Software also often has a preferred means of citing the software listed in the repository.

Are there any other non-software references, i.e. published papers, that are relevant?
