+++ paper/paper.bib
@@ -46,7 +46,8 @@
     title  = "recordr",
     author = "Peter Slaughter, Matthew B. Jones, Christopher Jones, Lauren Palmer (NCEAS)",
     url    = "https://github.com/NCEAS/recordr",
-    year   = "2020 (accessed June 30, 2020)"
+    year   = "2020 (accessed June 30, 2020)",
+    doi    = "doi:10.5063/F1GF0RF6"
 }
 @manual{recipy,
     title  = "recipy-Effortless method to record provenance in Python",
