+++ paper/paper.bib
@@ -36,11 +36,13 @@
     url    = "https://drivendata.github.io/cookiecutter-data-science/#cookiecutter-data-science",
     year   = "2020 (accessed June 30, 2020)"
 }
-@manual{sumatra,
-    title  = "Sumatra: automated tracking of scientific computations",
-    author = "Andrew Davison et al",
-    url    = "https://sumatra.readthedocs.io/en/latest/index.html",
-    year   = "2020 (accessed June 30, 2020)"
+@inbook{sumatra,
+    author = {Davison, Andrew and Mattioni, Michele and Samarkanov, D and Tele'nczuk, B},
+    year = {2014},
+    month = {03},
+    pages = {57-79},
+    title = {Sumatra: A Toolkit for Reproducible Research},
+    journal = {Implementing Reproducible Research}
 }
 @manual{recordr,
     title  = "recordr",
@@ -49,11 +51,29 @@
     year   = "2020 (accessed June 30, 2020)",
     doi    = "doi:10.5063/F1GF0RF6"
 }
-@manual{recipy,
-    title  = "recipy-Effortless method to record provenance in Python",
-    author = "",
-    url    = "https://github.com/recipy/recipy",
-    year   = "2020 (accessed June 30, 2020)"
+@misc{recipy,
+    author = {
+                Michael Jackson and
+                Robin Wilson and
+                Janneke van der Zwaan and
+                Daniel W. Steinbrook and
+                Florian Rathgeber and
+                Raquel Alegre and
+                Thomas Edwards and
+                Cash Costello and
+                Samuel Danny and
+                Jan Chwiejczak and
+                Michiel Cottaar and
+                Carlos Martinez-Ortiz and
+                Dave Vieglais and
+                Andrei Kucharavy and
+                Gary Ruben and
+                Utkarsh Upadhyay
+             },
+    title  = {recipy},
+    month  = {10},
+    year   = {2018},
+    url    = {https://github.com/recipy/recipy}
 }
 @manual{renv,
     title  = "Introduction to renv",
