### 2020-07-20 91e2c4d
Update references (#178)
* Add DOI for recordr

* Update recipy citation
