### 2020-07-20 #2456 @tamasgal
Alright, improved a bit but for the remaining packages I could not find better citations or papers. Those are mostly R packages...

Thanks for the hint with the Sumatra paper, I totally overlooked that.
