+++ paper/paper.bib
@@ -42,7 +42,8 @@
     month = {03},
     pages = {57-79},
     title = {Sumatra: A Toolkit for Reproducible Research},
-    journal = {Implementing Reproducible Research}
+    journal = {Implementing Reproducible Research},
+    doi = "doi:10.1201/9781315373461-3"
 }
 @manual{recordr,
     title  = "recordr",
