### 2020-07-22 50a088a
use relpath in collect_results
