### 2020-07-23 5da65a0
update docstring and changelog
