+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# 1.14.5
+* new keyword argument `rpath` for `collect_results!` that allows storing relative paths
 # 1.14.0
 * New macro `@onlyif` that allows placing restrictions on values in a dictionary used for expansion with `dict_list`
 # 1.13.1
