+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DrWatson"
 uuid = "634d3b9d-ee7a-5ddf-bec9-22491ea816e1"
 repo = "https://github.com/JuliaDynamics/DrWatson.jl.git"
-version = "1.14.5"
+version = "1.14.6"
 
 [deps]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
