### 2020-07-23 152c74e
don't use isnothing
