### 2020-08-14 7c328ac
Better README in initialize_project
That highlights reproducibility and instructs how to do it.
