### 2020-08-14 5bf3d95
Better README in initialize_project (#185)
That highlights reproducibility and instructs how to do it.
