### 2020-08-18 46137e8
Correct breakline for README enumeration (#187)
* Better README in initialize_project

That highlights reproducibility and instructs how to do it.

* correct breakline for enumeration of README
