### 2020-09-13 #2456 @Datseris
Hello,

I'd like to suggest a reviewer for our paper: James Fairbanks, PhD ( @jpfairbanks ) is familiar with the Julia language, is already familiar with various software for scientific assistanship, and is unaffiliated with DrWatson.

@dpsanders could you please consider him as a reviewer and possibly start the review process? I am meanwhile searching for a second reviewer as well.

best,
George
