### 2020-09-14 #2673 @whedon
[REVIEW]: DrWatson: the perfect sidekick for your scientific inquiries
**Submitting author:** @tamasgal (<a href="http://orcid.org/0000-0001-7821-8673">Tamás Gál</a>)
**Repository:** <a href="https://github.com/JuliaDynamics/DrWatson.jl" target ="_blank">https://github.com/JuliaDynamics/DrWatson.jl</a>
**Version:** 1.16.3

---
