+++ paper/paper.md
@@ -36,15 +36,15 @@ bibliography: paper.bib
 ---
 # Summary
 
-Doing scientific work always involves a lot of focus and scrutiny, since producing a scientific result requires several levels of depth of analysis, all of which must be as accurate and as reproducible as possible. 
-All this required scrutiny *should* be naturally translated into the codebase of the scientific project. 
+Doing scientific work always involves a lot of focus and scrutiny, since producing a scientific result requires several levels of depth of analysis, all of which must be as accurate and as reproducible as possible.
+All this required scrutiny *should* be naturally translated into the codebase of the scientific project.
 One should strive for a code that is doing what it is supposed to, it is reproducible, it doesn't break over time, it is sufficiently clear of bugs, simulation results are appropriately labelled, and more.
 The challenges associated with carrying out scientific work should not be made any worse by the difficulties of managing the codebase and resulting data/simulations.
 An unfortunate but likely outcome of this stress is that scientific codebases tend to be *sloppy*: folders are not organized, there is no version control, data are not provenanced properly, most scripts break over time, and the whole project is very hard, if not impossible, to reproduce.
 
 # Statement of need
 
-DrWatson is a **scientific project assistance software**. 
+DrWatson is a **scientific project assistance software**.
 Its purpose is to help scientists manage their scientific codebase in a simple and clear manner, to make the process of creating the codebase faster and to enable true full reproducibility and project sharing.
 DrWatson achieves this while being entirely non-invasive throughout the process.
 It provides a well-tested science-driven framework for managing a scientific project, thus removing the unnecessary stress and giving more time to doing actual science.
@@ -65,11 +65,11 @@ This means that DrWatson's features can be grouped into the following few main c
 * **Naming schemes**: A robust and deterministic scheme for naming and handling your data structures.
 * **Saving tools**: Tools for safely saving and loading your data, automatically tagging the Git commit ID to your saved files, and more.
 * **Running & listing simulations**: Tools for producing tables of existing simulations/results, adding new simulation results to the tables, preparing batch parameter containers, and more.
- 
+
 The next section illustrates these categories. For a more thorough explanation, the reader is referred to DrWatson's main documentation.
 
 # Typical workflow with DrWatson
-In this section we demonstrate how using DrWatson makes the typical scientific workflow faster, more robust, and easily reproducible. 
+In this section we demonstrate how using DrWatson makes the typical scientific workflow faster, more robust, and easily reproducible.
 This section is a brief summary of the [DrWatson Workflow Tutorial](https://juliadynamics.github.io/DrWatson.jl/dev/workflow/), which in itself showcases a subset of DrWatson's functionality.
 There the workflow is discussed and demonstrated more thoroughly via explicitly running every code command.
 
@@ -78,24 +78,24 @@ This creates a project folder that contains sensible default structure (e.g. fol
 This allows the scientific project to be tied with the full hierarchy of exact package versions used, which remains entirely independent from the main Julia installation (or any other project).
 The project is also a git repository, which allows code versioning and reproducibility, and DrWatson provides functions that make this process seamless (see below).
 
-Within the context of DrWatson, all project-related code runs *after* the corresponding Julia project has been activated. 
+Within the context of DrWatson, all project-related code runs *after* the corresponding Julia project has been activated.
 Several DrWatson functions like `projectdir`, `datadir`, `plotsdir` and similar are then made available.
 When these functions are called they always return the absolute path to the directory (or the appropriate subdirectories) in the active project, independently of the current working directory or the script directory these functions are called from.
[…]
