### 2020-09-14 #2673 @dpsanders
I would say no, that's up to the authors themselves.
