### 2020-09-15 JuliaDynamics/DrWatson.jl#190 @kescobo
Add test coverage metric
As part of [JOSS review](https://github.com/openjournals/joss-reviews/issues/2673), I'm asked to assess whether you have tests. Obviously, you do, and have CI + they pass locally on my machine, which is great, but it would be nice to see what your test coverage is as it's hard for me to gauge coming to the project for the first time.

Your tests are sufficient for me to check the box on the review, this is not essential, but it would be nice. There are examples all over the place for getting this to work with travis, but let me know if you're having trouble.
