### 2020-09-15 JuliaDynamics/DrWatson.jl#191 @kescobo
Add CONTRIBUTING.md or section to the docs
See "Community Guidelines" section of [JOSS review](https://github.com/openjournals/joss-reviews/issues/2673):

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
