### 2020-09-15 0d045ce
Create CONTRIBUTING.md
