+++ CONTRIBUTING.md
@@ -1,0 +1,12 @@
+# Contributing
+
+When contributing to this repository, consider first discussing the change you wish to make via an issue. 
+
+## Pull Request Process
+
+1. Ensure all tests pass locally before starting the pull request.
+1. Add adequate description in the Pull Request, or cite the corresponding issue if one exists by using hastag `#` and the issue number.
+1. Always allow the "editing from maintainers" option in your PR.
+2. Update the CHANGELOG.md with details of the (notable) changes to the exported API.
+3. Increase the version number in Project.toml according to [SemVer](http://semver.org/).
+4. Feel free to explicitly tag (with `@`) one of the main developers to request a review of your PR when it is ready.
