### 2020-09-15 JuliaDynamics/DrWatson.jl#191 @Datseris
Closed in https://github.com/JuliaDynamics/DrWatson.jl/commit/0d045ce77ad4d4bd97b607cce93c5ba88c616ae4
