### 2020-09-15 7d40e75
Paper review
Related to [JOSS review](https://github.com/openjournals/joss-reviews/issues/2673#event-3765935810)

I've made a few minor edits to the manuscript. A couple of additional points:

- I think it would also be good to mention in the `Comparison with existing software section`, where you discuss that many projects use multiple programming languages, that julia actually has really good interop with other languages. via eg PyCall and RCall.
- In the same section, I think cwl is orthogonal to the purpose of DrWatson. The idea with CWL and other pipeline-type software (see [this discourse
[…]
