### 2020-09-15 #2673 @kescobo
I've completed my review - pending the changes [requested here](https://github.com/JuliaDynamics/DrWatson.jl/pull/192), I'm happy to sign off.
