### 2020-09-16 #2673 @jpfairbanks
So I have a question about scope. Is the JOSS paper format intended to be a complete description of the code? How much of the "explaining how to use the software" should be delegated to the docs (which will stay up to date).
