### 2020-09-16 #2673 @dpsanders
Thanks @kescobo!

@jpfairbanks: That's a great question.
The JOSS paper is definitely *not* a complete description of the code.

There should be one or a few code snippets giving a feeling for how to use the code, to explain the goals etc. of the software.

The docs should have a complete API description etc.
