### 2020-09-16 #2673 @dpsanders
@jpfairbanks You can find review criteria here:

https://joss.readthedocs.io/en/latest/review_criteria.html
