### 2020-09-16 #2673 @dpsanders
The idea is for the paper to be relatively short, with the main documentation in the software repository itself.
