### 2020-09-16 #2673 @jpfairbanks
Thanks, I had read some of that language above in this thread and got confused because it talks about API docs, but I’m supposed to check that the API is documented in the actual manual. Not that the paper contains docs. This now makes sense. BTW I can’t mark the checkboxes in the checklist. Am I supposed to have different repo privs or something?
