### 2020-09-16 #2673 @dpsanders
OK, let me know if the wording can be improved.

I believe that you should have got an invitation to the JOSS account giving you those privileges?
