### 2020-09-16 b017c5b
WIP: Add coverage (#193)
* Add coverage

* add covecov badge to readme
