### 2020-09-16 JuliaDynamics/DrWatson.jl#190 @JonasIsensee
There is a codecov badge in the readme now.
