### 2020-09-16 JuliaDynamics/DrWatson.jl#190 @sebastianpech
Yeah pretty good, 100% is close 🤔. We're mostly missing warnings and errors.
