### 2020-09-16 700b2c2
use Normal package instead of command line
