### 2020-09-16 #2673 @Datseris
> Does JOSS have the concept of "corresponding author" and/or author order conventions?

For this paper you can consider myself the corresponding author (if need be).
