### 2020-09-16 3a956aa
Better address the comparison with CWL
