### 2020-09-16 #2673 @jpfairbanks
I think that this is a great tool and a great write-up that sufficiently introduces the software, situates it in the relevant prior art and motivates me to adopt the software based on claimed benefits. The benefits of better simulation data management is obvious to anyone who has analyzed the performance, accuracy, or predictions of a computer simulation. There is no empirical evidence of efficiency or utility presented, but I am not sure that publication of this material requires that kind of empirical validation and defer to the editor's judgement.

The biggest issue from a methodological perspective is that Dr. Watson avoids the use of databases for managing the data collected during the experiments. The current practice for simulation data management is to create files with paths that encode information about the parameters and with contents that encode information about the results. Dr. Watson automates and formalizes these ad-hoc techniques. In that automating and formalizing ad hoc techniques can improve the reliability of the methods, this is good. But as a field, computer science has solved this data management problem with an entirely different technique.

Relational Database Management Systems (RDBMS) are the standard technology designed for managing data when you need persistence and correctness with the ability to store data quickly and query against it later. RDBMSes support flexible query languages for answering complex, ad-hoc queries about the data and indexing to make it fast. Why does Dr. Watson not use a database for storing this data? The paper should address this approach and why it was not selected. The current approach for simulation data management is ad hoc file naming conventions and folder structures containing structured or semi-structured data in the output files. So any solution that starts by improving this solution is welcome. However for more fruitful analysis of the resulting data, a proper data ingestion with ETL into an SQL database would allow for more rich querying. This could lead to easier feature development down the road when it comes to experiment analysis.

This concern could be addressed by identifying previous attempts to use RDBMSes in simulation data management and explaining why they failed, or discussing this path as a viable approach that has not yet been explored, and leaving integration with existing RDBMS techniques for future work, once the RDBMS ecosystem in Julia has matured.
