### 2020-09-16 #2673 @Datseris
Thank you for the review @jpfairbanks , [CaosDB](https://caosdb.org/) is a RDBMS (I believe, although the creators name it Scientific Data Management System), and has been developed by scientists for scientists. We plan to integrate DrWatson with CaosDB.jl once a sufficiently stable Julia implementation exists. I believe once such an integration is in order, your concerns about the lack of data managing capabilities of DrWatson will be addressed. However, I do think this will take substantial amount of effort and thus it is unlikely that we are able to do it during the period of this review.
