### 2020-09-16 #2673 @jpfairbanks
That sounds great, then I would request that the article describe the limitation of DrWatson for organizing or querying the resulting data and that you would use DrWatson for tracking the data generation processes and then ETL it into CaosDB for querying. Describing that as future work would be appropriate.
