### 2020-09-16 #2673 @Datseris
Fair point, addressed in https://github.com/JuliaDynamics/DrWatson.jl/commit/bbe9a83dffc0e13d96285c8713db93ac1b695e89 .
