+++ paper/paper.bib
@@ -90,3 +90,17 @@
     url = "https://figshare.com/articles/dataset/Common_Workflow_Language_draft_3/3115156",
     doi = "10.6084/m9.figshare.3115156.v2"
 }
+
+@article{Fitschen2019,
+  doi = {10.3390/data4020083},
+  url = {https://doi.org/10.3390/data4020083},
+  year = {2019},
+  month = jun,
+  publisher = {{MDPI} {AG}},
+  volume = {4},
+  number = {2},
+  pages = {83},
+  author = {Timm Fitschen and Alexander Schlemmer and Daniel Hornung and Henrik tom W\"{o}rden and Ulrich Parlitz and Stefan Luther},
+  title = {{CaosDB}{\textemdash}Research Data Management for Complex,  Changing,  and Automated Research Workflows},
+  journal = {Data}
+}
