+++ paper/paper.md
@@ -99,6 +99,7 @@ into
 ```julia
 savename(parameters, "jld2")
 ```
+with output:
 ```
 "date=2020-05-23_pos_z=0.5_phi=3.jld2"
 ```
