### 2020-09-17 040a192
Update paper/paper.md
