### 2020-09-19 #2673 @jpfairbanks
> Please note that DrWatson is not a data management system and provides only basic data management functionality that remains self-contained in a single scientific project. One of our main future goals is to integrate DrWatson with a Relational Data Management System, specifically CaosDB (Fitschen et al., 2019), which had been developed specifically to handle large data bases connecting several scientific projects.
👍
