### 2020-09-20 0e42ae7
Add a "save" option to `savename`
