+++ src/naming.jl
@@ -15,7 +15,7 @@ The function chains keys and values into a string of the form:
 ```julia
 key1=val1_key2=val2_key3=val3
 ```
-while the keys are **always sorted alphabetically** by default. If you provide
+while the keys are **sorted alphabetically** by default. If you provide
 the prefix/suffix the function will do:
 ```julia
 prefix_key1=val1_key2=val2_key3=val3.suffix
@@ -59,8 +59,9 @@ See also [`parse_savename`](@ref) and [`@savename`](@ref).
   called with its default arguments (so customization here is possible only
   by rolling your own container type). If the `savename` of the nested
   containers is `""`, it is also skipped.
-* `sort = true` : Indicate whether the pairs are sorted alphabetically
-  by keys.
+* `sort = true` : Indicate whether the pairs are sorted alphabetically by
+  keys. If not, they are sorted by the order of `accesses`. WARNING: the 
+  default `accesses` is not deterministic for `Dict` inputs.
 
 ## Examples
 ```julia
