### 2020-09-21 b648030
Add a "sort" option to `savename` (#196)
* Add a "save" option to `savename`

* Documentation updates

* Update the version
