### 2020-09-21 #2673 @dpsanders
Thanks @apdavison for agreeing to review this as well. Dr. Andrew Davison is one of the main authors of the Python `sumatra` package.
