### 2020-10-06 0797a87
Fix dict_list wrongly promoting types (#199) (#200)
* Fix #199

* Bump version
