### 2020-10-22 #2673 @apdavison
@dpsanders I can't edit the checklist. When I click to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations it says the invitation has expired.
