### 2020-10-22 #2673 @apdavison
This seems like a nice toolkit for Julia users, which brings together a lot of functionality that would otherwise have to be assembled from several sources. I have the impression that a lot of the functionality relies on the modern design of Julia (in particular the package manager), and would be more difficult to implement in "messier" language ecosystems like Python's. The paper is well structured and clearly written.

I don't have any previous experience with Julia, so have not tried installing or using the package, but I think we can take the word of @kescobo and @jpfairbanks that this works!

My list of comments is not extensive, so I include it here rather than creating a new issue in the DrWatson repository.

- I don't think the "Community guidelines" item in the checklist has been well addressed. There is a CONTRIBUTING.md file, but as far as I can see the documentation itself doesn't contain "_guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_".
- Tele'nczuk should be Teleńczuk.
- Sumatra is not just for Python projects. It is written in Python but has support for R, MATLAB, bash, and is designed to be easy to extend for other languages, and to be usable in heterogeneous environments and pipelines.
- I think [ReproZip](https://www.reprozip.org) and [ActivePapers](https://www.activepapers.org) should be mentioned in the "Comparison" section, as tools which also aim to make scientific projects easily portable, but in more language-independent ways.
