### 2020-10-22 246eab2
Update CONTRIBUTING.md
