+++ paper/paper.md
@@ -165,10 +165,11 @@ While DrWatson comes with a predefined structure, packages like `rrtools` [@rrto
 Like DrWatson, most of the tools that initialize a folder structure, also initialize a Git repository for version control.
 In order to gain advantage from having code in version control eg. extracting diffs or commit ids, additional software packages, focused at data provenance, are needed.
 
-Applications like `sumatra` [@sumatra] for Python work mainly by executing scripts through a separate standalone tool that captures and tags all files created at runtime.
-Alternatives like `recordr` [@recordr] for R, `explore` for Matlab or `recipy` [@recipy] for Python aim at the non-invasive approach by redefining IO functions for logging metadata during saving.
+Applications like `sumatra` [@sumatra] which is written in Python and also supports MATLAB, R, BASH while it also provides extensibility for other languages,
+work mainly by executing scripts through a separate standalone tool that captures and tags all files created at runtime.
+Specialised alternatives like `recordr` [@recordr] for R, `explore` for Matlab or `recipy` [@recipy] for Python aim at the non-invasive approach by redefining IO functions for logging metadata during saving.
 The outlined tools, however, come with a cost of being limited to certain supported IO functions or the need of additional software to run code or a server infrastructure.
-Moreover, all of them are tied to a specific programming language and data provenance is only provided in their own context and usually within a single process.
+Moreover, most of them are tied to a specific programming language and data provenance is only provided in their own context and usually within a single process.
 Scientific projects, however, often deal with heterogeneous computing environments and pipelines running a multitude of scripts and applications connected to each other, thus the orchestration and data provenance needs to be implemented in a more language agnostic way.
 An example for such a framework (with significantly different end-goals compared to DrWatson) is the Common Workflow Language [@cwl].
 Notice that in principle DrWatson is tied to Julia, a single programming language. But because Julia has strong interop capabilities, allowing native C/FORTRAN calls and calls to Python or R (for example) via PyCall and RCall, the Julia-based design of DrWatson is much less of a limiting factor than for other languages.
