+++ paper/paper.md
@@ -166,7 +166,7 @@ Like DrWatson, most of the tools that initialize a folder structure, also initia
 In order to gain advantage from having code in version control eg. extracting diffs or commit ids, additional software packages, focused at data provenance, are needed.
 
 Applications like `sumatra` [@sumatra] which is written in Python and also supports MATLAB, R, BASH while it also provides extensibility for other languages,
-work mainly by executing scripts through a separate standalone tool that captures and tags all files created at runtime.
+work mainly by executing scripts through a separate standalone tool that captures and tags all files created at runtime. Another example for such an external manager is `ReproZip` [@ChirigatiRSF16] which traces system calls to identify which files are part of a specific analysis and generates additional metadata to combine everything together into zip-file in a reproducible manner.
 Specialised alternatives like `recordr` [@recordr] for R, `explore` for Matlab or `recipy` [@recipy] for Python aim at the non-invasive approach by redefining IO functions for logging metadata during saving.
 The outlined tools, however, come with a cost of being limited to certain supported IO functions or the need of additional software to run code or a server infrastructure.
 Moreover, most of them are tied to a specific programming language and data provenance is only provided in their own context and usually within a single process.
