+++ paper/paper.bib
@@ -104,3 +104,18 @@
   title = {{CaosDB}{\textemdash}Research Data Management for Complex,  Changing,  and Automated Research Workflows},
   journal = {Data}
 }
+
+@inproceedings{ChirigatiRSF16,
+  title = {ReproZip: Computational Reproducibility With Ease},
+  author = {Chirigati, Fernando and Rampin, R{\'e}mi and Shasha, Dennis and Freire, Juliana},
+  year = {2016},
+  isbn = {978-1-4503-3531-7},
+  location = {San Francisco, California, USA},
+  doi = {10.1145/2882903.2899401},
+  booktitle = {Proceedings of the 2016 International Conference on Management of Data},
+  series = {SIGMOD '16},
+  pages = {2085--2088},
+  numpages = {4},
+  publisher = {ACM},
+  keywords = {computational reproducibility, provenance, reprozip},
+}
