+++ paper/paper.bib
@@ -37,7 +37,7 @@
     year   = "2020 (accessed June 30, 2020)"
 }
 @inbook{sumatra,
-    author = {Davison, Andrew and Mattioni, Michele and Samarkanov, D and Tele'nczuk, B},
+    author = {Davison, Andrew and Mattioni, Michele and Samarkanov, D and Teleńczuk, B},
     year = {2014},
     month = {03},
     pages = {57-79},
