### 2020-10-22 #2673 @Datseris
Thank you @apdavison for the quick response and your review. We have addressed your comments by:

1. Adding a "Support & Contributing" section in the docs.
1. Improving the introduction of sumatra
1. Mentioning ReproZip (but not ActivePapers because contrary to its name it doesn't feel active enough of a repository to us)
1. Addressing typos
