### 2020-10-22 #2673 @dpsanders
@Datseris I agree with @apdavison that ActivePapers should be cited, even if it is not currently "active".
