+++ paper/paper.bib
@@ -119,3 +119,22 @@
   publisher = {ACM},
   keywords = {computational reproducibility, provenance, reprozip},
 }
+@article{HINSEN2011579,
+  title = "A data and code model for reproducible research and executable papers",
+  journal = "Procedia Computer Science",
+  volume = "4",
+  pages = "579 - 588",
+  year = "2011",
+  note = "Proceedings of the International Conference on Computational Science, ICCS 2011",
+  issn = "1877-0509",
+  doi = "https://doi.org/10.1016/j.procs.2011.04.061",
+  url = "http://www.sciencedirect.com/science/article/pii/S1877050911001190",
+  author = "Konrad Hinsen",
+  keywords = "Scientific computing, Scientific publications",
+  abstract = "This proposal describes how data, program code, and presentation can be stored together in a single file suitable for electronic publication and permitting the reproduction of computational results.. Universality, effciency, platformindependence, automated verifiability, and provenance tracking are the major design criteria. Existing and well-tested technology is used as much as possible, the two major building blocks being the Hierarchical Data Format for storage and the Java Virtual Machine for platform-independent code representation and secure execution."
+}
+@manual{activepapers,
+    title  = "ActivePapers",
+    url    = "http://www.activepapers.org/",
+    year   = "2020 (accessed October 22, 2020)"
+}
