### 2020-10-26 bfb25de
Minor edits
Minor edits for clarity and style.
