+++ paper/paper.md
@@ -38,11 +38,11 @@ bibliography: paper.bib
 
 Doing scientific work always involves a lot of focus and scrutiny, since producing a scientific result requires several levels of depth of analysis, all of which must be as accurate and as reproducible as possible.
 All this required scrutiny *should* be naturally translated into the codebase of the scientific project.
-One should strive for a code that is doing what it is supposed to, it is reproducible, it doesn't break over time, it is sufficiently clear of bugs, simulation results are appropriately labelled, and more.
+One should strive for a code that is doing what it is supposed to, is reproducible, doesn't break over time, is sufficiently clear of bugs, and with simulation results that are appropriately labelled, and more.
 The challenges associated with carrying out scientific work should not be made any worse by the difficulties of managing the codebase and resulting data/simulations.
 An unfortunate but likely outcome of this stress is that scientific codebases tend to be *sloppy*: folders are not organized, there is no version control, data are not provenanced properly, most scripts break over time, and the whole project is very hard, if not impossible, to reproduce.
-We've created the software DrWatson to make the process of scientific project management easier.
-In this paper we will describe how DrWatson results in an efficient scientific workflow, taking time off of project management and giving it to doing science.
+We have created the software DrWatson to make the process of scientific project management easier.
+In this paper we will describe how DrWatson results in an efficient scientific workflow, taking time away from project management and giving it to doing science.
 
 # Statement of need
 
@@ -53,10 +53,10 @@ It provides a well-tested science-driven framework for managing a scientific pro
 
 Technically DrWatson is a package for the Julia language.
 It is likely that Julia is the only language that can allow DrWatson to be the powerful framework that will be presented here, because Julia combines multiple dispatch, an integrated package manager, macros and code introspection.
-That said however, we believe that a large part of the design of DrWatson can be applied to other languages as well.
+That said, however, we believe that a large part of the design of DrWatson can be applied to other languages as well.
 
-In the rest of this paper we will provide a summary of what DrWatson can do (a detailed documentation that is regularly updated is hosted on [GitHub](https://juliadynamics.github.io/DrWatson.jl/stable/)).
-We will then highlight some examples of how using DrWatson speeds up the scientific workflow in real world scenarios.
+In this paper we provide a summary of what DrWatson can do (detailed documentation that is regularly updated is hosted on [GitHub](https://juliadynamics.github.io/DrWatson.jl/stable/)).
+We will then highlight some examples of how using DrWatson speeds up the scientific workflow in real-world scenarios.
 We close with a comparison with existing software.
 
 # Features and Functionality
@@ -71,7 +71,7 @@ This means that DrWatson's features can be grouped into the following few main c
 The next section illustrates these categories. For a more thorough explanation, the reader is referred to DrWatson's main documentation.
 
 Please note that DrWatson is not a data management system and provides only basic data management functionality that remains self-contained in a single scientific project.
-One of our main future goals is to integrate DrWatson with a Relational Data Management System, specifically CaosDB [@Fitschen2019], which had been developed specifically to handle large data bases connecting several scientific projects.
+One of our main future goals is to integrate DrWatson with a Relational Data Management System, specifically CaosDB [@Fitschen2019], which has been developed specifically to handle large data bases connecting several scientific projects.
 
 # Typical workflow with DrWatson
 In this section we demonstrate how using DrWatson makes the typical scientific workflow faster, more robust, and easily reproducible.
@@ -89,7 +89,7 @@ When these functions are called they always return the absolute path to the dire
 This establishes a relative-only path relationship within the project, which allows it to naturally run on other machines when shared or synced via e.g. a cloud service.
[…]
