### 2020-10-26 #2673 @dpsanders
@Datseris @tamasgal I sent a few minor corrections to the paper.

Once that's done I think the paper is ready to accept!
Please archive it at Zenodo and let me know the version and DOI. You need to make sure that the title and author info in Zenodo match the paper exactly. Thanks!
