### 2020-10-26 #2673 @Datseris
Hi @dpsanders thanks for the improvements to the paper.

Here is the Zenodo DOI: [10.5281/zenodo.4133570 ](https://zenodo.org/account/settings/github/repository/JuliaDynamics/DrWatson.jl#)

> You need to make sure that the title and author info in Zenodo match the paper exactly.

The only problem is, I don't know how to do this... I didn't find a way in the website to alter the archived Zenodo information (and Zenodo archives GitHub repositories automatically). Do I have to make a manual Zenodo entry?
