### 2020-10-26 #2673 @kescobo
@Datseris once you've linked the repo to Zenodo, every time you make a release (a github release, which may or may not happen automatically depending on your bots) it will automatically archive it
