### 2020-10-26 #2673 @Datseris
Yes I have done this. But David has some additional requirements regarding author info and paper title _in Zenodo_. I do not see this info somewhere at the moment...
