### 2020-10-26 #2673 @kescobo
Ahh, sorry I missed that. I think that you can manually edit the title and authors of your last release, and then it will automatically keep the same ones going forward. So:

1. Click "uploads" in the top banner
