### 2020-10-27 5de6670
some small improvements of the docs
