### 2020-10-27 #2673 @kescobo
> I'm now wondering if the title should be DrWatson.jl: the... instead of just DrWatson

`DrWatson` seems like the name of the software, `DrWatson.jl` is the package that implements it. I think either would work, but fine to leave as-is.
