+++ paper/paper.bib
@@ -43,14 +43,14 @@
     pages = {57-79},
     title = {Sumatra: A Toolkit for Reproducible Research},
     journal = {Implementing Reproducible Research},
-    doi = "doi:10.1201/9781315373461-3"
+    doi = "10.1201/9781315373461-3"
 }
 @manual{recordr,
     title  = "recordr",
     author = "Peter Slaughter, Matthew B. Jones, Christopher Jones, Lauren Palmer (NCEAS)",
     url    = "https://github.com/NCEAS/recordr",
     year   = "2020 (accessed June 30, 2020)",
-    doi    = "doi:10.5063/F1GF0RF6"
+    doi    = "10.5063/F1GF0RF6"
 }
 @misc{recipy,
     author = {
@@ -127,7 +127,7 @@
   year = "2011",
   note = "Proceedings of the International Conference on Computational Science, ICCS 2011",
   issn = "1877-0509",
-  doi = "https://doi.org/10.1016/j.procs.2011.04.061",
+  doi = "10.1016/j.procs.2011.04.061",
   url = "http://www.sciencedirect.com/science/article/pii/S1877050911001190",
   author = "Konrad Hinsen",
   keywords = "Scientific computing, Scientific publications",
