### 2020-10-29 #2673 @arfon
@Datseris - could you please merge this PR which fixes some issues with your DOIs? https://github.com/JuliaDynamics/DrWatson.jl/pull/208
